requirements.txt
asn1crypto==0.24.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
Django==1.11.15
django-filter==2.0.0
django-tinymce==2.7.0
djangorestframework==3.8.2
enum34==1.1.6
httpie==0.9.9
idna==2.7
ipaddress==1.0.22
Markdown==2.6.11
MySQL-python==1.2.5
Pillow==5.2.0
pkg-resources==0.0.0
pycparser==2.18
Pygments==2.2.0
PyMySQL==0.9.2
pytz==2018.5
requests==2.19.1
six==1.11.0
urllib3==1.23
uWSGI==2.0.17.1
